--- modules/container/src/main/resources/META-INF/mule-test-module.properties
@@ -1,3 +1,0 @@
-module.name=container
-
-artifact.export.classPackages=org.mule.runtime.container.internal
